+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added coverage for smart search, provided by Search-IMAsset.
+
+### Removed
+
+- Removed Get-IMAuditDelete as it is no longer provided by the Immich API.
+
 ## [1.118.0] - 2024-10-15
 
 ### Fixed